refactor: Standardize reservationId and rentalStatus type and fix params in navigation (#121)
- Changed `reservationId` type from `Long` to `Int` in `RentalChatRoomSummaryModel` and `RentalDto`.
- Updated `RentalDto.status` to use the `RentalStatus` enum instead of `String`.
- Simplified navigation to `ChatRoomScreen` in `RequestHistoryScreen` and `ChatListRoute` by removing redundant `productId` and `reservationId` parameters, relying solely on `chatRoomId`.
- Removed unnecessary logging for unknown rental statuses in `OwnerRentalDetailMapper` and `RenterRentalDetailMapper` as the status is now strongly typed.
- Adjusted sample data in `RenterRentalDetailScreen` and `OwnerRentalDetailScreen` to reflect the enum change for `status`.

+++ app/src/main/java/com/example/rentit/domain/rental/model/RentalChatRoomSummaryModel.kt
@@ -4,7 +4,7 @@ import com.example.rentit.common.enums.RentalStatus
 import java.time.LocalDate
 
 data class RentalChatRoomSummaryModel (
-    val reservationId: Long,
+    val reservationId: Int,
     val status: RentalStatus,
     val startDate: LocalDate,
     val endDate: LocalDate,
